Catch data on room

+++ domain/src/main/java/com/javalenciab90/models/Current.kt
@@ -5,7 +5,7 @@ data class Current(
     val feelslike: Int,
     val humidity: Int,
     val observation_time: String,
-    val precip: Int,
+    val precip: Double,
     val pressure: Int,
     val temperature: Int,
     val uv_index: Int,
